error handling and pydantic basemodels
