+++ static/styles.css
@@ -1014,9 +1014,6 @@ button {
     
 }
 
-.thankyou-container > body{
-}
-
 .thankyou-container > *{
     line-height: 20px;    
     margin-bottom: 24px;
