+++ static/thankyou.html
@@ -23,7 +23,7 @@
 
 <div class="thankyou-container body bold gray-70">
 <div class="header-3 cyan-70 bold">感謝您的訂購！</div>
-<div>您的訂單號碼是：<span class="thankyou-text cyan-70 btn bold"></span></div>
+<div>您的訂單號碼是：<span class="thankyou-text cyan-70 btn bold"><!--renders 訂單編號--></span></div>
 <div>此訂單已成功付款，請妥善保存此訂單號碼，謝謝！</div>
 <div>如有相關疑問，請寄信至tdt@example.com，並附上訂單號碼，將有專人為您處理。</div>
 </div>
